requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Web framework
Flask>=2.0.0
flask-cors>=3.0.10

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Data handling
joblib>=1.0.0
pyyaml>=5.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
pylint>=2.12.0

# Logging and utilities
python-dotenv>=0.19.0

# Optional: For production deployment
gunicorn>=20.1.0
